feat(numdigits): Add `numdigits.c` example
This example showcases the second loop in the book, which is a `do`
loop.

Whilst `while` loop can do essentially everything that `do` loop does,
it is sometimes preferred to use `do` loop for the cases when the actual
loop needs to run at least once.

This is illustrated in the example, since every number has at least one
digit, the loop body needs to run at least once for every single input
this program gets.

Signed-off-by: acikgozb <acikgozb@proton.me>