feat(length): Add `length.c` example
With this example, the author starts to dive in to different types such
as `char`.

The example showcases how to calculate the length of a given sentence by
using `getchar()` in combination with a loop.

It's a bit shady that we're not giving `getchar()` any buffer or
something to read from.

`man 3 getchar` answers this question, `getchar()` basically reads a
single character from `stdin`. The main function it uses can be seen
from the same man page.

Signed-off-by: acikgozb <acikgozb@proton.me>