feat(projects): Add Chapter 10 project solutions
For Chapter 10, the directory structure for the projects are changed.

At this point, since the projects keep getting a bit more complex
with each chapter, keeping all the solutions in a single file causes a
decrease in readability.

Therefore, this commit adds the solutions by creating new directories
for each of them.

Signed-off-by: acikgozb <acikgozb@proton.me>