feat(remind2): Add `remind2.c` example
This example showcases how dynamic memory allocation can be utilized by
using the previous `remind.c` example.

In the first version, we used a two-dimensional array to store all the
reminders.
The memory for all reminders in the `reminders` array were allocated by
the compiler in a fixed way.

In this version, we use a one-dimensional array to store the reminders.
The elements are basically null pointers.

Instead of allocating a fixed memory beforehand, we dynamically allocate
just enough memory for a single reminder as we insert it.

Signed-off-by: acikgozb <acikgozb@proton.me>