feat(inventory2): Add `inventory2.c` example
This example builds on the previous learnings about dynamic memory
allocation and demonstrates a perfect use case for such concept: linked
lists.

The author showcases a linked list implementation using `malloc` by
altering the `inventory.c` example that he wrote in Chapter 16.

In this version, instead of having a fixed size for the inventory, he
changes the main data structure of an inventory (array) to a linked
list. By doing this, the program no longer limits the user from entering
a part (as long as there is enough memory to store one).

The example is complex in a way that not only he introduces linked list,
the insertions done on the linked list is done in a sorted way (by part
number).

Signed-off-by: acikgozb <acikgozb@proton.me>