feat(stack): Add `stack` example
This example showcases how a C program can be modularized.

`stack.h` header file (similar to the idea of traits/interfaces in other
programming languages) can be used to store the declarations that can be
used by the "clients".

The files `stack1.c` and `stack2.c` correpsond to different
implementations that define the functions declared in the header file.

By using the `static` keyword in implementations, we can prevent
clients from accessing the implementation details and alter the
functionality of a module.

So, a client can include `stack.h` and based on the linked
implementation, they can use either `stack1.c` or `stack2.c`.

Signed-off-by: acikgozb <acikgozb@proton.me>