feat(stack-adt): Add `stackADT.c` example
This example showcases the usage of incomplete types in C, which is the
only way (so far) we have to create abstractions around types.

In this example, an abstract data type is created for a stack data
structure, and its fields are kept hidden from the client.

The example utilizes this abstract data type (ADT) and shows three
different implementations of a stack:

- Via a fixed size array.
- Via a dynamic size array.
- Via a linked list.

Signed-off-by: acikgozb <acikgozb@proton.me>