feat(invclear): Add `invclear.c` example
The author showcases the usage of block I/O and file positioning in this
example by reading from/writing to a binary file.

The binary file in the example is basically a memory block of an
`inventory`, which is an array of `part` structs.

Since the original binary file is not provided in the example, I
modified it a little bit to make it generate 2 binary files.

This makes it a bit easier to understand how we modify the file via
`fwrite` and `rewind`.

Signed-off-by: acikgozb <acikgozb@proton.me>